requirements.txt
pytorch_lightning
matplotlib
numpy
scipy
Pillow
rasterio
albumentations
segmentation_models_pytorch

